backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
Pillow==10.4.0
python-jose[cryptography]==3.3.0
requests==2.32.3
pymongo==4.6.0
motor==3.3.2
python-geohash==0.8.5
python-dotenv==1.1.1

